+++ src/main/kotlin/dev/ocpd/jsensible/rules/jpa/JpaRules.kt
@@ -1,10 +1,8 @@
 package dev.ocpd.jsensible.rules.jpa
 
 import com.tngtech.archunit.lang.ArchRule
-import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses
 import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noMembers
 import dev.ocpd.jsensible.internal.jpa.EagerFetch.useEagerFetch
-import dev.ocpd.jsensible.internal.jpa.NotSpringNullable.notUseSpringNullable
 
 /**
  * General Java/Jakarta Persistence API (JPA) rules.
@@ -12,8 +10,7 @@ import dev.ocpd.jsensible.internal.jpa.NotSpringNullable.notUseSpringNullable
 object JpaRules {
 
     fun all() = listOf(
-        noEagerFetch(),
-        onlyUseSpringNullableAnnotation()
+        noEagerFetch()
     )
 
     /**
@@ -32,19 +29,4 @@ object JpaRules {
         noMembers().should(useEagerFetch())
             .`as`("use eager fetch")
             .because("no property should be fetched eagerly by default")
-
-    /**
-     * Only support the [org.springframework.lang.Nullable] annotation
-     *
-     * Spring Data Repositories only support the usage of their own
-     * [org.springframework.lang.Nullable] annotation. Misuses could
-     * lead to runtime errors when invoking repository methods.
-     *
-     * Solution: Use the [org.springframework.lang.Nullable] annotation.
-     */
-    fun onlyUseSpringNullableAnnotation(): ArchRule =
-        noClasses()
-            .that().areAssignableTo("org.springframework.data.repository.Repository")
-            .should(notUseSpringNullable())
-            .because("only support the [org.springframework.lang.Nullable] annotation]")
 }
